+++ BixiOnHBase/src/org/apache/hadoop/hbase/coprocessor/BixiProtocol.java
@@ -26,13 +26,13 @@ public interface BixiProtocol extends CoprocessorProtocol {
   Map<String, Integer> giveAverageUsage(List<String> stationIds, Scan scan)
       throws IOException;
 
-  Map<String, Double> getAvailableBikesFromAPoint(double lat, double lon,
+  Map<String, Integer> getAvailableBikesFromAPoint(double lat, double lon,
       double radius, Get get) throws IOException;
   
   Map<String, Integer> getTotalUsage_Schema2(List<String> stationIds, Scan scan)
   throws IOException;
 
-  Map<String, Double> getAvailableBikesFromAPoint_Schema2(double lat, double lon,
+  Map<String, Integer> getAvailableBikesFromAPoint_Schema2(double lat, double lon,
 		  Scan scan) throws IOException;
   List<String> getStationsNearPoint_Schema2(double lat, double lon) throws IOException;
